Cleanup dashboard layout and move images to the assets folder
