+++ console/src/components/MapCore/MapCore.js
@@ -11,7 +11,7 @@ import ri from '@enact/ui/resolution';
 import AppContextConnect from '../../App/AppContextConnect';
 import appConfig from '../../App/configLoader';
 import {propTypeLatLon, propTypeLatLonList} from '../../data/proptypes';
-import CarPng from '../Dashboard/svg/car.png';
+import CarPng from '../../../assets/car.png';
 import {ServiceLayerContext} from '../../data/ServiceLayer';
 
 import css from './MapCore.less';
