+++ console/src/views/Dashboard.less
@@ -11,7 +11,7 @@
 	.progressBar {
 		display: inline-block;
 		margin: 0 3px;
-		width: 200px;
+		width: 240px;
 	}
 
 	.sideInfo {
